agents/multi-agents/requirements.txt
python-dotenv
smolagents[litellm]
matplotlib
geopandas
shapely
kaleido